# openapi.yaml
openapi: 3.0.3
info:
  title: LoadFlow API
  version: 1.3.0
  description: |
    Welcome to the LoadFlow API!
    
    This API provides a complete backend for the LoadFlow Dispatch Application, allowing dispatchers, drivers, and administrators to orchestrate logistics seamlessly.

    ### 🔑 Authentication

    The system uses **Cookie-based JWT Authentication**.

    1. **Login** or **Register** (if invited) to receive a secure HTTP-only cookie.
    2. This cookie is automatically sent with subsequent requests.
    3. **Note for API Tools (Postman/Swagger):** Ensure cookies are enabled/stored after login so protected endpoints can be tested.

    ### 📦 Primary Features
    
    This API is built explicitly to handle fleet operations.
    
    - **Logistics Data:** Access resources in JSON format.
    - **Driver Dashboard Metrics:** Real-time calculation of Service Hours, Miles, and Completed Missions.
    - **Smart Notifications:** Optimistic state updates and ultra-fast polling for instant driver/dispatcher feedback.
    - **Proof of Delivery:** Secure `multipart/form-data` uploads handled and routed towards Cloudinary.
    - **Server Actions:** Public gateways directly dispatch SMTP NodeMailer notifications without persistence overhead.

    ### 👥 User Roles

    - **Driver:** Can view assigned configurations, update transit statuses, track performance metrics (Miles/Hours), and submit Proof of Delivery.
...
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: Get users
      description: Available to Admin and Dispatcher roles
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get a user by id
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Users]
      summary: Update a user
      description: Admin only
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Version conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Users]
      summary: Delete a user
      description: Admin only
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/users/{id}/unlock:
    post:
      tags: [Users]
      summary: Unlock a locked account
      description: Admin only
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Account unlocked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/{id}/revoke:
    post:
      tags: [Users]
      summary: Revoke all active sessions for a user
      description: Admin only. Increments the target user's token version.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Sessions revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: MongoDB document id
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        _id:
          type: string
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [Admin, Dispatcher, Driver]
        isPending:
          type: boolean
        phone:
          type: string
          nullable: true
        licenseNumber:
          type: string
          nullable: true
        dob:
          type: string
          format: date-time
          nullable: true
        city:
          type: string
          nullable: true
        province:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        resetPasswordRequested:
          type: boolean
        resetPasswordApproved:
          type: boolean
        loginAttempts:
          type: integer
        isLocked:
          type: boolean
        lockedUntil:
          type: string
          format: date-time
          nullable: true
        tokenVersion:
          type: integer
          description: Session revocation index
        isTwoFactorEnabled:
          type: boolean
          description: Whether the user has setup 2FA
        requestedBy:
          type: string
          nullable: true
          description: Dispatcher ID for recruitment flow reference
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        __v:
          type: integer
...
    UserUpdateRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [Admin, Dispatcher, Driver]
        isPending:
          type: boolean
        phone:
          type: string
        licenseNumber:
          type: string
        city:
          type: string
        province:
          type: string
        postalCode:
          type: string
        address:
          type: string
        resetPasswordRequested:
          type: boolean
        resetPasswordApproved:
          type: boolean
        loginAttempts:
          type: integer
        isLocked:
          type: boolean
        __v:
          type: integer
...